+++ projects/car-zone/service/interface.go
@@ -6,12 +6,12 @@ import (
 )
 
 type Car interface {
-	GetCarById(ctx context.Context, id int64) (*models.Car, error)
+	GetCarById(ctx context.Context, id int) (*models.Car, error)
 	GetCarByBrand(ctx context.Context, brand string) ([]*models.Car, error)
 	CreateCar(ctx context.Context, carReq *models.CarRequest) (*models.Car, error)
 }
 
 type Engine interface {
-	GetEngineById(ctx context.Context, id int64) (*models.Engine, error)
+	GetEngineById(ctx context.Context, id int) (*models.Engine, error)
 	CreateEngine(ctx context.Context, engineReq *models.EngineRequest) (*models.Engine, error)
 }
